requirements.txt
pandas
numpy
scikit-learn
google-generativeai
Flask
